Revert "Ignore the water thickness head in the channel model"
This reverts commit ae981817616596be19c52fad5a9089161d9c8d63.

It seems like we would want the channel model to see the elevation head
produced by a lake's water thickness to ensure that channel flow is
consistent with the presence of the lake.  This commit reverts the
commit that dropped the water thickness elevation head from the
hydropotential gradients used by the channel.

In the sheet model, the elevation head is removed from the advection
equation, but it is handled by the diffusion equation, so it's still
represented in the evolution.  As the channel model currently exists,
the elevation head is almost completely absent (other than its effect
on the magnitude of gradMagPhiEdge).  It seems like this could lead to
channels draining *into* lakes even when the full hydropotential
gradient of the lake indicates water flow should be draining the lake.