Response to review comments - add ability to deep copy interp fields.
This commit responds to reviewer comments.  The biggest change is to
allow for a deep copy of the fields that will be used for interpolation.
The user is now able to specify that rather than having the time interpolation
work with a clone of a specific field it can deal with the field directly.
This will remove storing an extra copy of a field that may not be needed.
Additionally, it will save on the deep copy that would be needed to transfer
data from the time interpolation clone to the field in the simulation.

Other changes are to remove the shift_data_impl and just have one shift_data
call.  Fix a typo in the comments.