Treat tangent slope calculation near boundary of mesh
This commit explicitly treats the calculation of hydropotentialBaseSlopeTangent
near mesh boundaries.
For the two vertex based methods, edges with vertices on the boundary
of the mesh will have invalid values.
For the normal slope method, any edges on cells that are at the
edge of the mesh will have contaminated values.
In those cases, set tangent slope to zero rather than leaving the
possibility of garbage values.