Channel gradMagPhiEdge dependent on full water hydropotential
Channels now use updated version of gradMagPhiEdge, which is dependent
on full hydropotential. This is done to keep channels from filling up
lakes even when lake is full and  hydropotential is in opposite direction.

Distributed system is now dependent on new variable gradMagPhiBaseEdge,
which is dependent only on hydropotentialBase and is equivalent to
gradMagPhiEdge in previous commits.