Add iceThicknessHydro to replace thickness
thickness field is replaced by iceThicknessHydro, which makes changes
along the boundaries where the surface evelation of a boundary cell
is a local minimum. In this case the surface elevation of that cell is
replaced by the minimum of its non-boundary neighbors, and the ice
thickness is then recalculated from the updated surface elevation.

The intent of this commit is to inhibit the formation of subglacial
lakes forming on the domain boundaries, which causes instabilities in
the channel model. Physically, lakes should not form at the boundary as
it is intended to be a high in hydropotential (a watershed boundary).