Updated timing of lnd data average for the iac
The annual average of lnd data for the iac is now performed
during the setup send call immediately before the iac run call.
This is ensure proper restart values, although it appears that the
values should be identical to the previous timing of the average
call. The problem appears to be non-bit-for-bit calculations in
less than 1% of records that are passed to the iac, which causes
small deviations in these scalars from run to run (also <1%).
Possibly rounding the values could avoid this.