iceThicknessHydro additionally adjusts elev. highs
'iceThicknessHydro' now adjusts the ice thicknesses of boundary cells
so that there are also no local maxima in surface elevation on the
boundaries (in addition to the previous removal of local minima).
This will make the surface elevation of all boundary
cells equal to either the minimum or maximum of their non-boundary
neighbors. This is done to avoid diffusivity instabilities when using
the 'from_vertex_barycentric' tangent slope calculation.

Currently, this implementation will change *all* grounded boundary
cells, so this commit also introduces a new namelist option,
'config_SGH_use_iceThicknessHydro' that enables the 'iceThicknessHydro'
calculation if set to '.true.'. 'iceThicknessHydro' is equivalent to 'thickness' when
'config_SGH_use_iceThicknessHydro' is false.
'config_SGH_use_iceThicknessHydro' has a default value of '.true.'