Replace masked values with nearest unmasked values in nudging targets.
This commit scans every target nudging state for masked values that
would be put there by the vertical interpolation routine.  When a masked
value is found it is replaced with the nearest unmasked value in the column
following these rules:
1. If the masked value is near the top-of-model and there are no unmasked
   values above it then the masked value is assigned the nearest unmasked
   value "below" it.
2. The algorithm then scans from top to bottom until it hits a masked value,
   if it does then it assigns that level in the column the last unmasked
   value it saw.
This has the effect of top filling to the nearest unmasked value near TOM and
backfilling at the surface to the nearest unmasked value closest to the surface.