Add capability to use a static set of source pressure levels in nudging.
This commit gives the option to the user to use a static set of pressure
levels for the nudging source data rather than have the nudging process
look for a `p_mid` variable in the source data.  This is useful for cases
where the source data may have been generated from an EAMxx vertical interpolation
output onto set pressure levels.

In order to facilitate this change this commit also had to make it an option
to vertically interpolate when the x_src is 1D.  This was introduced in the last
commit but had to be fixed a bit here.