Add a get_variable_metadata function
This commit introduces the get_variable_metadata function to the IO
interface.  This will be useful for gathering information about a variable
during a run, which can be used for checks, etc.

In this commit the getter only is defined for float and double, but the
interface is general enough to allow for a char based getter to be defined
in the future.  The getter is needed to identify the Fill Value used by a
variable which will be checked in the nudging process.