Add support for masked/filled values in scorpio_output
This commit enhances the `combine` function in scorpio_output.cpp
to support cases where some values may be masked or filled.  This
is essential to properly handle output that utilizes vertical remapping,
which may include some filled values.

There is a different treatment of filled values depending on the averaging
type for the output.
 - INSTANT: no change
 - MIN and MAX: Only use the new_val to update if it is unfilled.
 - AVERAGE: Only add the contribution of the new_val if it is unfilled.

*Note on AVERAGE.  Because the AVERAGE output is later scaled by the
number of points that contribute to the average this commit also tracks
the number of times a new value was used to update the output value over
the averaging period.  This is then used with AVERAGE to scale the output.