Add a threshold for when averaging output will be "filled"
This commit adds a runtime option to specify an averaging coefficient
threshold that is used to determine if an output point should be filled,
when the averaging type is AVERAGE.  The threshold compares the number of
unfilled contributions against the number of snaps before the last write.

If that ratio is greater than the threshold then the point will take the
average of the unfilled contributions, otherwise the point will be assigned
the fill value.

The default threshold is 0.5 or 50%