Fix incorrect wind "addition" in gustiness calculations.
As described in Ma et. al. (2022), the correct equation for combining
two sources of gustiness with the resolved wind speed is

U^2 = U0^2 + a UZM^2 + b UCLUBB^2

where U is the desired gustiness-enhanced wind, and

 * U0 is resolved wind.
 * UZM is deep convective gustiness.
 * UCLUBB is turbulent/boundary-layer gustiness.
 * a and b are tuning parameters.

The prior implementation did not follow this formula (see #5835),
by failing to properly account for the squaring of each wind speed
in the formula (or in the case of UCLUBB, accounting for it
incorrectly). This commit fixes that bug.

Note that the parameters "a" and "b" also no longer multiply the
diagnostic outputs 'VMAGDP' and 'VMAGCL'. This results in output that
is more consistent with the literature, though less consistent with
previous model versions. Multiplying by sqrt(a) and sqrt(b) can
easily be done in post-production, however.

This commit also allows atm_gustiness to be used with downscaling in
ELM; previously, these features were not allowed to be combined, but
that restriction appears to have been unnecessary.

Ma et. al. (2002): https://doi.org/10.5194/gmd-15-2881-2022

Reported-by: @mabrunke

[CC] for all configurations with EAM, but should be bit-for-bit for
all other configurations. Expected effect is a reduction in the
effect of gustiness, resulting in larger resolved wind speeds but
smaller surface fluxes where the resolved wind speed is still low.