Prevent ELM from adding extra gustiness when EAM supplies gustiness.
When EAM is running, it provides a gustiness value that includes the
effects of boundary layer turbulence. This is redundant with the
calculation of convective gustiness `wc` that is performed by ELM.
See #5837. This commit turns off the calculation of `wc` when the
`atm_gustiness` namelist variable is set, signifying that EAM is
providing the gustiness value.

This commit also adds the `force_land_gustiness` option to ELM. This
option forces the old behavior, where ELM adds gustiness regardless
of what the atmosphere model does.

[CC] for cases where EAM is running, should be bit-for-bit otherwise.
Expected effect is to increase wind speeds over land, and likely
reduce surface fluxes a bit when wind speed is low.
[NML]