Create a separate combine function allowing for mask values
This commit makes a fully separate version of combine that is used for
cases when a mask value is provided.  This is a different strategy than
the lest commit which attempted to overload the single combine function using
a default argument for the mask value.  This led to a compiler error.