Make conserve tracer volume
Fix a bug in fct tracer advection that caused tracers to not be conserved
due to an inconsistent treatment of the thickness flux. The previous
implementation inadvertently used first-order upwind thickness flux
to calculate higher-order tracer flux, which led to loss of conservation
when both thickness and tracers used fct. Note that this still does
not conserve tracers to machine precision, even though mass is conserved
to machine precision. But testing shows the tracer volume conserved
to within about 0.02%.