Chrysalis: Exclude chr-0512 for one test.
A good lead in my nondeterminism hunt on Chrysalis is that node chr-0512
produces the failure very frequently, while other nodes I have tried so far
never do. Since the small_kernels test happens to fail the most, probably
because of how nodes get allocated to the tests, I want to see what happens when
it isn't permitted to use this node.