Response to review: Fix hard coded fill values in field and combine funcs
This commit addresses the reviewer comment that we are hard-coding the fill value
in combine_and_fill, update, update_impl and scale to be of the type `float`.
The rest of the arguments for these functions are templated.  This commit fixes that,
now fill_val is either not passed or, if it is, it is passed as a templated type.

This commit also fixes the surface_coupling unit test to reflect changes in an
earlier commit (the one dealing with alternate names in surface coupling).