EAMxx: add ability to modify atm proc group on the fly in atmchange
The syntax is

```
  $ ./atmchange group_name=ap_1[,ap_2..]
```
where the ap_i's describe the new atm proc group. There are two possibilities:
- if ap_i is in the xml defaults, buildnml will simply grab it
- if ap_i is NOT in the xml defaults, atmchange will add a new group with
  that name in the defaults (*)

If a new atm proc group is added, it is added with no atm procs inside,
which will cause an error at runtime. The user is responsible of doing
```
$ ./atmchange ap_i=<list of procs>
```
afterwards, in order to make the new group non-empty.

Note: if the user makes a typo such as
```
  $ ./atmchange physics=shoc,cld_fraction,p3,rtmgp
```
(notice the misspelling of rrtmgp), the framework will not detect that,
and will assume that rtmgp is a new group.

(*) the new atm proc group is NOT written to the xml defaults file. It is
just added to the in-memory XML tree, for the duration of the buildnml call