Add non-dynamic cells to openOceanMask to fix calving bug
Add non-dynamic cells to openOceanMask to fix bug that shut down
calving when there was a non-dynamic cell between the dynamic margin
and the open ocean. This is slightly inconsistent with the variable
name "openOceanMask", but the other implementation option involved
looping over neighbors of neighbors, which would be messier and more
expensive. This also should fix the edge case in which a single non-
dynamic cell shuts down calving in an entire embayment between two
areas of floating ice, which could happen at Thwaites, for instance.