remove redundant waterPressure GL alteration
removes condition that keeps grounding line waterPressure from dropping
below waterPressure in ocean. This is redundant now that ocean
waterPressure is zero.

THIS COMMIT MAY NEED TO BE REVERTED ONCE WE SETTLE ON TACTIC FOR
HANDLING WATER PRESSURES IN OCEAN, BUT I'M COMMITING THIS FOR NOW TO
COMPLETE REBASE ON DEVELOP - AOH 5/30/24