This commit addreses a current issue with add_field(U or V)
In the current setup, the fields U and V are added to the field manager
as subfields of horiz_winds *after* the field manager has already registerd
all field requests from atmospheric processes.  As a result, nudging, which
may register U and V disrupts the assumptions about horiz_winds currently
hard-coded.  When U and V are used in nudging they are added as independent
fields and are no longer tied to horiz_winds.

This commit is a quick fix to address the immediate issue in nudging.  A longer
term fix will be require a bit more work and will be added in a subsequent PR.