Refactor blas and lapack stuff a bit more
Get rid of all the BLA_VENDOR stuff in the macros. It had no impact
because deps are loading before macros are processed. We should prefer
using the environment to set this stuff like we are doing for the other
libs. pm-cpu_nvidia seems to be the one platform sensitive to lapack/blas
so can be used as an example of how to specify these libraries.