PIO: Remove dependence on CIME/Tools/Makefile
We need to remove all dependencies on CIME/Tools/Makefile so
we can change our macros to use cmake-style names for things.
PIO/Scorpio was the first thing I picked.

The situation with spio was strange. It was using CIME/Tools/Makefile
to call CMake. It translated the Makefile-style settings to Cmake using
a lot of extra code in the Makefile to set up CMAKE_OPTS.

We change the python wrapper build script, buildlib.spio, to instead
just set up the cmake options in python, allowing us to skip the CIME/Tools/Makefile.

On top of removing the dependency, I think this makes the overall spio
build a lot easier to understand, although it was a bit tedious pulling
things out of the macros.

Once the macros are in proper cmake-stype, I think we could just pass them
directly to scorpio's CMake build via -C and things will be even simpler.

[BFB]