EAMxx: fix calculation of tendencies
When an atm proc is not subcycled, but its part of an atm
proc group that is subcycled, we cannot overwrite what is
already in the tend field with the current atm proc step
tendency. The only (or, at least, simplest) way out is to
store the beg-of-step field in a separate copy.