Removing redundant env var WITH_HDF5_SCORPIO in buildlib.spio
If HDF5_ROOT is already set as an env var, it is sufficient to turn
on the WITH_HDF5 CMake option for scorpio. Introducing a new env var
like WITH_HDF5_SCORPIO may lead to unnecessary confusion.

Additionally, for now, maintain the WITH_HDF5 CMake option as OFF by
default to prevent potential build errors with the default scorpio
1.4.2 submodule.