version that passes caar_ut test for ORIGINAL and !ORIGINAL versions
the test was run with disabled TESTER_NOMPI in cmake for
both options, define ORIGINAL and undef ORIGINAL, to test that we did not break
the original code and to test that the inlined version is also bfb with it.

vvdp array is disabled in !ORIGINAL version for now, because udp is used in other
functors, so udp should remain defined in m_buffers.

besides there were minor clean ups.