Completed CO2 downscaling
The model successfully runs to completion for ZLND  with functional CO2
downscaling. The linearly downscaled CO2 data have been verified
within the GCAM repo code, but not yet in the fortran ehc code
(giac repo). Convergence downscaling has not been verified. Restarts
work properly. The gcam namelist has been updated to add several
options for convergence downscaling and the gcam co2 passing namelist
item has been corrected in name. A shell script has been added to
gcam/tools that generates the new input base co2 gridded files.
The fortran floating point exception checks are still off because
the model still fails due to untrapped GCAM nans if these are on
(which usually is only when DEBUG=TRUE). This allows for DEBUG to
be set to TRUE and for the model to run. The configuration for
the boost library for the ehc/gcam now points to the files in
the gcam repo for portability. The xerces library configuration
has been removed as it is no longer needed by GCAM.  The three repo
branches (e3sm, giac, gcam) at this commit work together
(at least for ZLND), and thse branches are set in the .gitmodules files.