EAMxx: pass number of physical levs to trcmix
We cannot rely on pmid/q for this info, since one (or both) of them
may be padded. Since we don't know if they are both padded, we may
as well pass nlevs.

NOTE: I thought about setting nlevs to the min of pmid/q.extent(1),
but this may still do extra/unnecessary work, since the loop is not
packed.