Add option to output a field at height above the surface.
The current implementation of FieldAtHeight assumes the user wants
to output the field at a specific height above sealevel.  In practice
this means the reference height used for interpolation is either z_mid
or z_int.

It is likely that users would also like the option to output a field at
a specific height above the surface, for example the 2 meter temperature.

This commit adds that capability by including a reference surface to the
FieldAtHeight diagnostic.  This can either be "sealevel" or "surface".
If a user specifies "sealevel" then z_mid or z_int are used as the level
heights.  If a user specifies "surface" than geopotential_mid or
geopotential_int are used instead.

This is controlled by the diagnostic name.  The old way to trigger a
FieldAtHeight was to list the variable X at N meters like follows:
  X_at_Nm
The new approach allows for
  X_at_Nm_above_sealevel, or
  X_at_Nm_above_surface.

If a user still uses the old way without specifying "above_Y" then it
defaults to "sealevel" and prints a warning to the ATM log.