Cleaned up code and fixed cice history bug
Compiling C++ libraries is now conditional on EHC being active
(along with a couple other conditional active components).
The number of pfts and harvest cats are still hardcoded in the
coupler (but cleaner) because the proposed namelist read was
dependent on something else.
Comments stating that EHC must be on same grid as EAM have been
removed.
The CICE history files exists flag is now broadcast across procs
because history writes were failing due to different flag values,
which caused the entire code to fail.