Add grounded ice mask to `externalWaterInput` term
The grounded ice mask is used in the calculation of the `basalMeltInput`
term but was excluded from `externalWaterInput`, which could cause
problems in closing budgets. The grounded ice mask was not added to
any of the calculations dependent on `waterThickness` b/c SGH model
explicitly sets the `waterThickness` to zero outside of the grounded
ice area.