Modify how dimensions ids are set in coupler history
The coupler history was not using the dimension id of coordinate variables
when defining dimensions of variables for most of the attribute vectors. This made
the coupler history hard to use with viz/analysis tools like ncvis. Allow
dimension ids of coordinate variables to be passed back to calling function
and then used in future calls that define additional variables.  Coupler history
can then be viewed with ncvis and possibly other tools.