intermediate try to configure co2_cycle with gcam interactions
created some temporary infrastructure to store the co2 variables,
the interpolation has to happen here.
modified the readflux_fuel and readflux_aircraft nl controls to
account for ehc being present to feed data into these paths
instead of reading the files.
currently failing on init when adding constituents. there should
only be 4, but it is saying that there are more than pcnst=40.