Implement the MAMDryDep::run_impl() function.
This is an interface to the mam4::DryDeposition::compute_tendencies()
function. This means taking Scream Kokkos views of data and
mapping them to mam4 Kokkos views of data. Most of the time
it is just a sub-view so memory is not created or destroyed.
There were a couple of choices for mapping which Scream
data to the mam4 data that need to be reviewed.

The resulting tendencies are in a local Kokkos view
called tendencies. These need to be made available to Scream.

This needs to be compiled and tested on a GPU.