Point to correct, updated glsmask file for GIS 20km configurations.
Previously, this pointed to a mask where glc was allowed to be active
everywhere. This change points to the correct mask and correct mask name
that has been updated so that GIS footprint describing the mask was made
from the current GIS 20km init. cond. file. Testing confirms the correct
mask is pulled from the repo and the relevant test cases run to
completion as expected when using this mask file.