Add a precipitation adjustment to the ml-correction process.
This commit adds a precipitation adjustment in cases where the ml-correction
is applied to qv.  We interpret the hard adjustment of qv in the column as
an attempt by ML to represent the missing precipitation in the coarse resolution
run vs. the fine resolution run.  In order to have balance with the land models
we need to add this adjustment to the precipitation, otherwise we risk drying out
the land.