Make changes in ml_correction thread safe, address review comments.
This commit adds Kokkos::single(Kokkos::PerTeam... to the conditionals
in the new precip adjustment.

This commit fixes an error in how qv_told was defined, which was originally
just an alias but should now be a truly separate copy of the initial qv state.

This commit also addresses review comments about naming of new variables.