Adding path to the timing bin dir
Adding path to the timing binary directory to the set of include
directories. This change is required for these tests that need
path to the fortran perf mod.

This fixes build issues with the HOMMEBFB test suite using the
latest version of the SCORPIO library (new versions of the SCORPIO
library no longer export internal SCORPIO build paths with the
SCORPIO library targets)