Fix indexing of layers in high freq. output
Previously, the layer above the one containing the desired
depth was being selected, rather than the layer containing the
depth.

The default depth if no layer is found is now the deepest layer,
rather than the first layer.  This is because, if no layer is found,
it means that even the deepest layer is shallower than the desired
depth, and the only sensible default is the deepest layer.