Add FOLISIO compset
This commit adds a FOLISIO compset (First-Order Land Ice, Sea Ice,
Ocean) that uses the Albany First-Order velocity solver in MALI instead
of the standard shallow-ice approximation solver.  The SIA solver is
inappropriate for Antarctica and was causing CFL violations when used
with the new mesh, making it not practical to use even for smoketesting.