Fix bottomDepth when filling bathymetry holes
In MPAS-Ocean init mode, when filling holes in the bathymetry
(cells with deepest layers that are deeper than any of their
neighbors) as part of the Haney-number-constrained vertical
coordinate used for ice shelves, the bottom depth was not also
being updated correctly for some cells.  This merge fixes the
issue for cells in the z-level region of the ocean (far from
ice-shelf cavities) by setting the bottom depth to the deepest
depth within the same z level as the deepest neighbor.