Disable hipInit call before MPI_Init on Frontier
The hipInit workaround was originally introduced to prevent
occasional segmentation faults during MPI_Init. With the
cpe/22.12 configuration in the post-maintenance setup on
Frontier, this workaround is no longer necessary.

Additionally, it seems that the hipInit call may be causing
non-deterministic DIRK NaN errors in recent ne1024 DECADAL
runs. So far, no DIRK errors have been observed when hipInit
is not invoked.