EAMxx: make rrtmgp call find_package for kokkos
If an existing install of Kokkos is used, the imported
target has local scope, so following calls to cmake's
target_link_libraries won't recognize Kokkos if the
first call to find_package was in a subdir.

In EAMxx, that's the case (it's found while parsing the
ekat subdir). However, subsequent calls to find_package
are virtually no-op, so it doesn't hurt to call it again