Add config_glc_thermal_forcing_coupling_mode option
This commit adds an MPAS-Ocean namelist option for activating the
OCN-GLC TF coupling.  This option controls if E3SM should enable the
OCN-GLC TF coupling and also makes the associated TF calculations
conditional on the option being active.  The option defaults to false
and there currently are not any compsets that activate it, so it can
only be enabled with a namelist usermod at present.