fix sat hist capability and add test
Ressurecting a long-forgotten capability to output
variable along ship/flight tracks, and adding a
test to (weakly try to) ensure it doesn't break again.

The code is rewritten such that it uses a different
IO logic (more serial) and support is added for
different types of output variables.

Fixes #6543

[BFB]

Note: I don't have full provenance from where the
code edits are from. I recall attempting to fix this
capability in July and August 2024, and I recall
trying different versions of available CAM code to do so.
But the final version isn't easily traceable
(so it could be a mix of edits from CAM and debug edits
to make it work in current EAM). For reference, here's
the CAM code:
https://github.com/ESCOMP/CAM/blob/cam_cesm2_1_rel/src/control/sat_hist.F90.

I don't claim ownership of the code at all, and I am
happy to give "authorship" to whomever claims it.