EAMxx: changes for print style in atmquery
Before we had two styles, 'short', and 'full'. The former only
printed the leaf name, the latter printed the full scope.

I changed those to 'node-name' and 'full-scope', and added a new
style called 'parent-scope', which only adds one parent to the
scope of the name. These types are used by default for:

- node-name: when printing recursively
- full-scope: when using --grep (unless the grep matches a non-leaf
  node, in which case the nested values are printed with node-name format)
- parent-scope: the default when printing a parameter