Fixes a bug when copying grid data for a geo data of scalar type.
When copying the grid geo data from one grid to the another there is
a check if the geo data has COL or not.  This led to an error for the
unique case of a geo data that is a scalar and thus the layout is
essentially empty.

This happens for DPxx cases.