EAMxx: create rrtmgp allsky baseline target with CreateUnitTest utility
The cmake utility links (indirectly) against scream_io, which avoids
link errors in case of shared libraries. The errors are due to the
fact that scream_share does use scream_io stuff, but the rrtmgp
executable was not linking against it. Since the exec was indeed
not using ANY of the scream_io symbols, static linking is fine, but
dynamic linking is not.