First step towards building E3SM on the LLNL machine tuolumne.
This commit adds a new machine config entry for tuolumne,
Adds a cmake_macro for tuolumne, and
Adds a cmake file for building EAMxx on tuolumne.

Testing has shown the E3SM will build, but we have not yet successfully
run E3SM on tuolumne.  Note, tuolumne is in early access at the moment.

Further work will need to be done to confirm that all modules are correct and
to actually finish a run.