Add variable ocn z-level coupler support
This commit adds the ability to support a variable number of ocean
z-levels for coupling ocean state fields between OCN and GLC.  The
intended use case is passing profiles of thermal forcing (and possibly
eventually temperature and salinity) from OCN to GLC, for MALI to use
for driving parameterizations of marine melting that is unresolved in
MPAS-Ocean (facemelting or ice shelves that are unresolved on the ocean
mesh).  The implementation closely mirrors that already existing for GLC
multiple elevation classes used for coupling surface mass balance with
the LND component.  This commit introduces the changes to the coupler
and configuration files needed to support multiple z-ocean classes, but
no changes are made to introduce them into MPAS-Ocean or MALI.