Add 3d TF support to MPAS-Ocean
This commit adds Registry options and MPAS-Ocean code to support a
time-averaged 3d thermal forcing field.  This follows the implementation
of a 2d thermal forcing field at a single depth that was added recently.
For now, both 2d and 3d TF functionalities are supported.

The time-averaged 3d TF values are calculated in a new variable,
avgThermalForcingAtZLevels.  The z-levels at which to calculate TF are
stored in a variable, glcZLevels.  The size of glcZLevels is defined by
a new dimension, nGlcZLevels, which is set by a new namelist option,
config_n_glc_z_levels.  The existing namelist option
config_glc_thermal_forcing_coupling_mode is modified to support a ‘3d’
option.  mpas_ocn_time_average_coupled.F is modified to calculate
avgThermalForcingAtZLevels if the 3d calculation is enabled.

Note that in this commit, there is no functionality to initialize
glcZLevels.  This will happen in a subsequent commit through the
coupler.  Also note that a package should probably be added for the new
fields.