Don't use extrapolation in OCN2GLC_TF_SMAPNAME mapping file
For the extrapolation in MALI to work properly, extrapolation in the
OCN2GLC_TF_SMAPNAME mapping file needs to be disabled.  As an initial
test, this commit switched from the bilin (which includes extrap) to the
aave (which does not) mapping file.  The correct behavior in MALI now
occurs.  A later commit will need to add bilin mapping files without
extrapolation for the OCN2GLC_TF_SMAPNAME mapping file for each mesh
that includes one.