Move MALI marine melting from rofl to rofi coupling stream
This way it will also extract latent heat from the ocean.  Note that
doing it this way, it will be combined with the calving flux.  Long
term, we will likely want to handle them separately so they can use
different horizontal and vertical distributions.  But at present MALI's
calving flux is not even hooked up, so that will need to be addressed at
a later date.