Add 3d TF masking
This commit adds a 3d mask field to MPAS-Ocean for tracking what
fraction of time steps in the time average are valid.  If the fraction
exceeds 0.9, then the driver considers that location valid and passes
the value to the coupler along with a coupler mask value of 1.  This new
coupler mask field matches the dimensionality of the 3d TF field and is
passed to MALI.  MALI evaluates the mask in its import routine and only
uses TF from locations that have at least 50% overlap with MPAS-Ocean
cells.  In those locations, it uses the TF passed through the coupler,
scaled by the area fraction.  In location that don't satisfy that
criterion, the MALI extrapolation mask is set to 0 and an invalid TF
value is inserted in the 3d TF field.