Remove layer just below bedTopography from availOceanMask
Remove layer directly below the bed topography in the ocean from the
availOceanMask and ensure it is not used in the interpolating to
calculate TF below ice shelves. Otherwise, these cells are assigned a
value of 0 degrees. Additionally, remove interpolation of 3d thermal
forcing above and below the bed, and instead calculate 2d thermal
forcing just from water properties in cell just above the bed.