Add OCN2GLC_TF_SMAPNAME filenames
These are appropriate for the in-progress 3d thermal forcing coupling and not
appropriate for the current 2d thermal forcing coupling.  They are set
to the esmfaave map files because those do not include extrapolation,
whereas the esmfbilin do, and it is important the OCN2GLC_TF_SMAPNAME
maps do not include extrapolation.  These may be replaced by esmfbilin
versions *without* extarpolation as part of the 3d TF coupling PR, but
the files specified here provide functional maps until then.