Accumulate calvingThickness from restore-calving in RK loop
Accumulate calvingThickness from restore-calving in RK loop, which was
previously being zeroed out when calving was called after RK loop.
This does not fully close the mass budget, but it reduces the error
considerably. Also, it is likely that calvingThickness will be zeroed
out again if using restore-calving along with a rate-based calving law,
so more work is likely needed.