Disable 3d TF coupling for MALI DATA and STATIC modes
We want 3d TF coupling on for all compsets with both MPAS-Ocean and MALI
active because we want to use 3d TF for facemelting even if we use
ice-shelf basal melt fluxes from the coupler.  However, we do not want
facemelting when MALI is DATA or STATIC modes, so the TF coupling should
be inactive in those cases.