Restore HDF5_ROOT ENV variable on Frontier for E3SM build
On Frontier, CMake's built-in FindHDF5 module fails with compilers
like craycray-mphipcc due to reliance on compiler wrapper tools that
may be unavailable or incompatible.

To resolve this, we provide a simplified custom FindHDF5.cmake that
avoids using compiler wrappers. It uses the HDF5_ROOT environment
variable to probe the installation and find the necessary libraries.

With this change, HDF5_ROOT can be enabled again on Frontier without
introducing build issues. Additionally, there is no longer a need to
set HDF5_USE_STATIC_LIBRARIES in FindPIO.cmake.