namelist variable for water flux conservation in ocean-only simulation
A section in MPASO namelist to make it flexible to change the tuning parameters of precipitation factor:
&sfwf_forcing
 config_precip_factor = 1.0
 config_sfwf_balance = .true.
 config_sfwf_balance_type = 0
 config_sfwf_history_days = 366
/
config_precip_factor: is the value of the tunnable factors applied to the snow and rain fluxes to maintain the freshwater balance in the G-Case simulation.

config_sfwf_balance:  a flag that was added to turn on (= .true.) or off (= .false.) this option

config_sfwf_balance_type: second flag variable to setup the options (default = 0 meaning that simply use the value of config_precip_factor to scaling down or up the surface freshwater fluxes (rain+snow) )

config_sfwf_history_days: currently is tentative and not used when config_sfwf_balance_type = 0

Currently, the config_sfwf_balance_type = 0 has been tested with the v3.LR model code, and I can verify it worked as expected.