Add code for tracking the minimum daily beta factor (btran).
The proposed pull request pretty much imports the code from CLM5 to compute the minimum
daily beta factor (btran). This is currently a diagnostic variable that is more sensitive
to water stress than the daily/monthly averages of btran. I am revisiting the drought
deciduous phenology in FATES, and this is a candidate for driving the abscission and
flushing times.