EAMxx: fix bug in horizontal remap
We cannot look at the smallest GID in the map file
to figure out the base index for GIDs in the map file,
since the map need not be surjective or total, meaning
that there may be elements in the codomain or domain
that are not mapped (to or from) by the map.

To fix this, we accept a map file "base_gid" when building
the remapper data. For now, default to 1, which is the
typical value for ncremap files