Add ability to use data SMB for MALI in a G-case
This commit adds the ability for MALI to use a data surface mass balance
field from the MALI input file in compsets where ELM is not active.
This is done by adding a new namelist option in MALI called
config_smb_source and making its value controlled by the CIME variable
GLC_TWO_WAY_COUPLING which is only true when both ELM and MALI are
active.