Remove bed topography dependence from 200-500 param
Removes dependence of the 200-500 TF parameterization on bed topography
to be consistent with original implementation. Parameterization assumes
no knowledge of fjord geometry, so water properties should be
extrapolated all cells below sea level and averaged between 200-500m,
regardless of whether this depth is below ocean bathymetry (all vertical
levels are valid ocean cells using this parameterization). This commit also
adds an error if the 200-500 parameterization is used with the sill
blocking extrapolation designation, as these are mutually exclusive.